refactor(engine): use integers for event/record length
Use `int` type instead of `long` for events/records. Under the hood, we
always use `int` and not `long`, simply widening the values to return
longs. It seemed unnecessary in the end, especially since we always deal
with the `ByteBuffer` type in the end, which only allows buffers with up
to `Integer.MAX_VALUE` length.